+++ CHANGELOG.md
@@ -146,6 +146,8 @@ Releases are done everytime an important ticket is completed.
 * **2021-08-05**:
    1. Rename `net.splitcells.dem.utils.CommonFunctions#hash_code` to `hashCode`.
 ### Minor Changes
+* **2021-12-20**: Create command `ssh.execute.in.background` in order to start
+  background tasks on remote servers with systemd and ssh.
 * **2021-12-19**: **\#138** Create rating cache for constraints via
   ConstraintAspect.
   In the future, it may be implemented in such a way,